@@ -25,11 +25,13 @@ dependencies:
     sdk: flutter
 
   get: ^4.3.8
-  bloc: ^7.0.0  
-  flutter_bloc: ^7.0.0
+  bloc: ^7.2.1
+  dartz: ^0.10.0
+  equatable: ^2.0.3
+  flutter_bloc: ^7.3.1
   flutter_hooks: ^0.18.0
-  shared_preferences: ^2.0.7
-  font_awesome_flutter: ^9.1.0
+  shared_preferences: ^2.0.8
+  font_awesome_flutter: ^9.2.0
   smooth_page_indicator: ^1.0.0+2
 
 
