@@ -6,12 +6,12 @@ import 'space_widget.dart';
 
 class CustomButton extends StatelessWidget {
   final String label;
-  final Color? color;
+  final Color? buttonColor;
   final VoidCallback onTap;
   const CustomButton({
     required this.onTap,
     required this.label,
-    this.color = kMainColor,
+    this.buttonColor = kMainColor,
     Key? key,
   }) : super(key: key);
 
@@ -23,7 +23,7 @@ class CustomButton extends StatelessWidget {
           width: SizeConfig.safeBlockHorizontal * 100,
           decoration: BoxDecoration(
             borderRadius: BorderRadius.circular(8.0),
-            color: color,
+            color: buttonColor,
           ),
           child: Center(
             child: Text(
@@ -44,14 +44,14 @@ class CustomButtonWithIcon extends StatelessWidget {
   const CustomButtonWithIcon({
     required this.onTap,
     required this.iconData,
-    required this.color,
+    required this.iconColor,
     this.text = 'Log in with',
     Key? key,
   }) : super(key: key);
   final String text;
   final IconData iconData;
   final VoidCallback onTap;
-  final Color color;
+  final Color iconColor;
   @override
   Widget build(BuildContext context) => GestureDetector(
         onTap: onTap,
@@ -72,14 +72,14 @@ class CustomButtonWithIcon extends StatelessWidget {
                 text,
                 style: const TextStyle(
                   fontWeight: FontWeight.w600,
-                  fontSize: 12.5,
+                  fontSize: 13,
                 ),
                 textAlign: TextAlign.left,
               ),
               const SpaceWidget(widthSpace: 0.5),
               Icon(
                 iconData,
-                color: color,
+                color: iconColor,
               ),
             ],
           ),
